add power spectrum visualization

+++ model/gpt2cgan.py
@@ -26,7 +26,7 @@ class gpt2cgan(tf.keras.Model):
         # add one hot vector for every seed
         self.seed = tf.random.normal([generated_count, noise_len, noise_dim])
 
-        tmp = [0] * (generated_count / 2) + [1] * (generated_count / 2)
+        tmp = [0] * int(generated_count / 2) + [1] * int(generated_count / 2)
         l = tf.constant(tmp)
 
         # l = tf.constant([x % 2 for x in range(2)])
